Jitter refactor
