+++ src/plugins/intel_gpu/src/graph/impls/ocl_v2/utils/kernel_generator.cpp
@@ -6,7 +6,7 @@
 
 #include <cctype>
 
-#include "common_utils/jitter.hpp"
+#include "jitter.hpp"
 #include "common_utils/kernel_generator_base.hpp"
 #include "intel_gpu/graph/program.hpp"
 #include "intel_gpu/runtime/kernel_args.hpp"
@@ -116,7 +116,6 @@ std::string KernelGenerator::get_build_options(const RuntimeParams& params) cons
 JitConstants KernelGenerator::get_jit_constants(const RuntimeParams& params) const {
     auto jit = make_base_jit_constants(params);
     jit.add(make_tensors_jit_constants(params));
-    jit.add(make_activation_jit_constants(activation_func::none, ov::element::dynamic, "", false, false));
     return jit;
 }
